requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0
numpy>=1.23.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
